+++ src/js/lib/underscore.mixins.js
@@ -9,8 +9,9 @@
 
 
 ( // Module boilerplate to support browser globals, node.js and AMD.
-  (typeof module !== "undefined" && function (m) { module.exports = m(); }) ||
-  (typeof define === "function" && function (m) { define('underscoreDeepExtend', m); }) ||
+  // node.js and AMD disabled for now
+  // (typeof module !== "undefined" && function (m) { module.exports = m(); }) ||
+  // (typeof define === "function" && function (m) { define('underscoreDeepExtend', m); }) ||
   (function (m) { window['underscoreDeepExtend'] = m(); })
 )(function () { return function(_) {